chore: Process Event as a separate workflow, and add an ollama workflow
