+++ app/actions.tsx
@@ -6,12 +6,9 @@ import {
   getAIState,
   getMutableAIState
 } from 'ai/rsc'
-import { CoreMessage, generateId, ToolResultPart } from 'ai'
-import { Spinner } from '@/components/ui/spinner'
+import { CoreMessage, generateId } from 'ai'
 import { Section } from '@/components/section'
 import { FollowupPanel } from '@/components/followup-panel'
-import { inquire, researcher, taskManager, querySuggestor } from '@/lib/agents'
-import { writer } from '@/lib/agents/writer'
 import { saveChat } from '@/lib/actions/chat'
 import { Chat } from '@/lib/types'
 import { AIMessage } from '@/lib/types'
@@ -21,9 +18,8 @@ import SearchRelated from '@/components/search-related'
 import { CopilotDisplay } from '@/components/copilot-display'
 import RetrieveSection from '@/components/retrieve-section'
 import { VideoSearchSection } from '@/components/video-search-section'
-import { transformToolMessages } from '@/lib/utils'
 import { AnswerSection } from '@/components/answer-section'
-import { ErrorCard } from '@/components/error-card'
+import { workflow } from '@/lib/actions/workflow'
 
 async function submit(
   formData?: FormData,
@@ -51,9 +47,6 @@ async function submit(
       const { role, content } = message
       return { role, content } as CoreMessage
     })
-
-  // groupId is used to group the messages for collapse
-  const groupId = generateId()
 
   const useSpecificAPI = process.env.USE_SPECIFIC_API_FOR_WRITER === 'true'
   const useOllamaProvider = !!(
@@ -100,174 +93,13 @@ async function submit(
     })
   }
 
-  async function processEvents() {
-    // Show the spinner
-    uiStream.append(<Spinner />)
-
-    let action = { object: { next: 'proceed' } }
-    // If the user skips the task, we proceed to the search
-    if (!skip) action = (await taskManager(messages)) ?? action
-
-    if (action.object.next === 'inquire') {
-      // Generate inquiry
-      const inquiry = await inquire(uiStream, messages)
-      uiStream.done()
-      isGenerating.done()
-      isCollapsed.done(false)
-      aiState.done({
-        ...aiState.get(),
-        messages: [
-          ...aiState.get().messages,
-          {
-            id: generateId(),
-            role: 'assistant',
-            content: `inquiry: ${inquiry?.question}`,
-            type: 'inquiry'
-          }
-        ]
-      })
-      return
-    }
-
-    // Set the collapsed state to true
-    isCollapsed.done(true)
-
-    //  Generate the answer
-    let answer = ''
-    let stopReason = ''
-    let toolOutputs: ToolResultPart[] = []
-    let errorOccurred = false
-
-    const streamText = createStreamableValue<string>()
-
-    // If ANTHROPIC_API_KEY is set, update the UI with the answer
-    // If not, update the UI with a div
-    if (process.env.ANTHROPIC_API_KEY) {
-      uiStream.update(
-        <AnswerSection result={streamText.value} hasHeader={false} />
-      )
-    } else {
-      uiStream.update(<div />)
-    }
-
-    // If useSpecificAPI is enabled, only function calls will be made
-    // If not using a tool, this model generates the answer
-    while (
-      useSpecificAPI
-        ? toolOutputs.length === 0 && answer.length === 0 && !errorOccurred
-        : (stopReason !== 'stop' || answer.length === 0) && !errorOccurred
-    ) {
-      // Search the web and generate the answer
-      const { fullResponse, hasError, toolResponses, finishReason } =
-        await researcher(uiStream, streamText, messages)
-      stopReason = finishReason || ''
-      answer = fullResponse
-      toolOutputs = toolResponses
-      errorOccurred = hasError
-
-      if (toolOutputs.length > 0) {
-        toolOutputs.map(output => {
-          aiState.update({
-            ...aiState.get(),
-            messages: [
-              ...aiState.get().messages,
-              {
-                id: groupId,
-                role: 'tool',
-                content: JSON.stringify(output.result),
-                name: output.toolName,
-                type: 'tool'
-              }
-            ]
-          })
-        })
-      }
-    }
-
-    // If useSpecificAPI is enabled, generate the answer using the specific model
-    if (useSpecificAPI && answer.length === 0 && !errorOccurred) {
-      // modify the messages to be used by the specific model
-      const modifiedMessages = transformToolMessages(messages)
-      const latestMessages = modifiedMessages.slice(maxMessages * -1)
-      const { response, hasError } = await writer(uiStream, latestMessages)
-      answer = response
-      errorOccurred = hasError
-      messages.push({
-        role: 'assistant',
-        content: answer
-      })
-    }
-
-    if (!errorOccurred) {
-      const useGoogleProvider = process.env.GOOGLE_GENERATIVE_AI_API_KEY
-      const useOllamaProvider = !!(
-        process.env.OLLAMA_MODEL && process.env.OLLAMA_BASE_URL
-      )
-      let processedMessages = messages
-      // If using Google provider, we need to modify the messages
-      if (useGoogleProvider) {
-        processedMessages = transformToolMessages(messages)
-      }
-      if (useOllamaProvider) {
-        processedMessages = [{ role: 'assistant', content: answer }]
-      }
-
-      streamText.done()
-      aiState.update({
-        ...aiState.get(),
-        messages: [
-          ...aiState.get().messages,
-          {
-            id: groupId,
-            role: 'assistant',
-            content: answer,
-            type: 'answer'
-          }
-        ]
-      })
-
-      // Generate related queries
-      const relatedQueries = await querySuggestor(uiStream, processedMessages)
-      // Add follow-up panel
-      uiStream.append(
-        <Section title="Follow-up">
-          <FollowupPanel />
-        </Section>
-      )
-
-      aiState.done({
-        ...aiState.get(),
-        messages: [
-          ...aiState.get().messages,
-          {
-            id: groupId,
-            role: 'assistant',
-            content: JSON.stringify(relatedQueries),
-            type: 'related'
-          },
-          {
-            id: groupId,
-            role: 'assistant',
-            content: 'followup',
-            type: 'followup'
-          }
-        ]
-      })
-    } else {
-      aiState.done(aiState.get())
-      streamText.done()
-      uiStream.append(
-        <ErrorCard
-          errorMessage={answer || 'An error occurred. Please try again.'}
-        />
-      )
-    }
-
-    isGenerating.done(false)
-    uiStream.done()
-  }
-
-  processEvents()
+  // Run the agent workflow
+  workflow(
+    { uiStream, isCollapsed, isGenerating },
+    aiState,
+    messages,
+    skip ?? false
+  )
 
   return {
     id: generateId(),
@@ -359,10 +191,10 @@ export const getUIStateFromAIState = (aiState: Chat) => {
   const chatId = aiState.chatId
   const isSharePage = aiState.isSharePage
 
-    // Ensure messages is an array of plain objects
-    const messages = Array.isArray(aiState.messages) 
-    ? aiState.messages.map(msg => ({...msg})) 
-    : [];
+  // Ensure messages is an array of plain objects
+  const messages = Array.isArray(aiState.messages)
+    ? aiState.messages.map(msg => ({ ...msg }))
+    : []
 
   return messages
     .map((message, index) => {
