+++ lib/agents/query-suggestor.tsx
@@ -25,14 +25,6 @@ export async function querySuggestor(
 
     For instance, if the original query was "Starship's third test flight key milestones", your output should follow this format:
 
-    "{
-      "related": [
-        "What were the primary objectives achieved during Starship's third test flight?",
-        "What factors contributed to the ultimate outcome of Starship's third test flight?",
-        "How will the results of the third test flight influence SpaceX's future development plans for Starship?"
-      ]
-    }"
-
     Aim to create queries that progressively delve into more specific aspects, implications, or adjacent topics related to the initial query. The goal is to anticipate the user's potential information needs and guide them towards a more comprehensive understanding of the subject matter.
     Please match the language of the response to the user's language.`,
     messages: lastMessages,
