+++ lib/agents/tools/search.tsx
@@ -41,7 +41,6 @@ export const searchTool = ({ uiStream, fullResponse }: ToolProps) =>
       const searchAPI =
         (process.env.SEARCH_API as 'tavily' | 'exa' | 'searxng') || 'tavily'
 
-     
       const effectiveSearchDepth =
         searchAPI === 'searxng' &&
         process.env.SEARXNG_DEFAULT_DEPTH === 'advanced'
@@ -55,7 +54,8 @@ export const searchTool = ({ uiStream, fullResponse }: ToolProps) =>
       try {
         if (searchAPI === 'searxng' && effectiveSearchDepth === 'advanced') {
           // API route for advanced SearXNG search
-          const baseUrl = process.env.NEXT_PUBLIC_BASE_URL || 'http://localhost:3000';
+          const baseUrl =
+            process.env.NEXT_PUBLIC_BASE_URL || 'http://localhost:3000'
           const response = await fetch(`${baseUrl}/api/advanced-search`, {
             method: 'POST',
             headers: { 'Content-Type': 'application/json' },
